auth-service/requirements.txt.txt
flask
pyjwt
flask_cors